Move tests to calamari folder
