+++ runtime/calamari/Cargo.toml
@@ -109,6 +109,46 @@ substrate-wasm-builder = { git = 'https://github.com/paritytech/substrate.git', 
 reqwest = { version = "0.11", features = ["blocking"] }
 serde_json = "1.0"
 
+# 3rd dependencies
+codec = { package = "parity-scale-codec", version = "3.1.2" }
+scale-info = { version = "2.1.2", features = ["derive"] }
+
+# Substrate dependencies
+frame-system = { git = 'https://github.com/paritytech/substrate.git', branch = "polkadot-v0.9.26" }
+pallet-assets = { git = 'https://github.com/paritytech/substrate.git', branch = "polkadot-v0.9.26" }
+pallet-balances = { git = 'https://github.com/paritytech/substrate.git', branch = "polkadot-v0.9.26" }
+pallet-utility = { git = 'https://github.com/paritytech/substrate.git', branch = "polkadot-v0.9.26" }
+sp-core = { git = 'https://github.com/paritytech/substrate.git', branch = "polkadot-v0.9.26" }
+sp-io = { git = 'https://github.com/paritytech/substrate.git', branch = "polkadot-v0.9.26" }
+sp-std = { git = 'https://github.com/paritytech/substrate.git', branch = "polkadot-v0.9.26" }
+
+# Polkadot dependencies
+pallet-xcm = { git = 'https://github.com/paritytech/polkadot.git', branch = "release-v0.9.26" }
+polkadot-core-primitives = { git = 'https://github.com/paritytech/polkadot.git', branch = "release-v0.9.26" }
+polkadot-parachain = { git = 'https://github.com/paritytech/polkadot.git', branch = "release-v0.9.26" }
+polkadot-runtime-parachains = { git = 'https://github.com/paritytech/polkadot.git', branch = "release-v0.9.26" }
+xcm = { git = 'https://github.com/paritytech/polkadot.git', branch = "release-v0.9.26" }
+xcm-builder = { git = 'https://github.com/paritytech/polkadot.git', branch = "release-v0.9.26" }
+xcm-executor = { git = 'https://github.com/paritytech/polkadot.git', branch = "release-v0.9.26" }
+xcm-simulator = { git = 'https://github.com/paritytech/polkadot.git', branch = "release-v0.9.26" }
+
+# Cumulus dependencies
+cumulus-pallet-dmp-queue = { git = 'https://github.com/paritytech/cumulus.git', branch = "polkadot-v0.9.26" }
+cumulus-pallet-parachain-system = { git = 'https://github.com/paritytech/cumulus.git', branch = "polkadot-v0.9.26" }
+cumulus-pallet-xcm = { git = 'https://github.com/paritytech/cumulus.git', branch = "polkadot-v0.9.26" }
+cumulus-pallet-xcmp-queue = { git = 'https://github.com/paritytech/cumulus.git', branch = "polkadot-v0.9.26" }
+cumulus-primitives-parachain-inherent = { git = "https://github.com/paritytech/cumulus.git", branch = "polkadot-v0.9.26" }
+cumulus-test-relay-sproof-builder = { git = "https://github.com/paritytech/cumulus.git", branch = "polkadot-v0.9.26" }
+parachain-info = { git = 'https://github.com/paritytech/cumulus.git', branch = "polkadot-v0.9.26" }
+
+# Orml dependencies
+orml-traits = { git = "https://github.com/manta-network/open-runtime-module-library.git", branch = "polkadot-v0.9.26" }
+orml-xtokens = { git = "https://github.com/manta-network/open-runtime-module-library.git", branch = "polkadot-v0.9.26" }
+
+# Self dependencies
+pallet-asset-manager = { path = '../../pallets/asset-manager' }
+
+
 [features]
 default = ['std']
 runtime-benchmarks = [
