+++ runtime/calamari/tests/integrations_mock/mock.rs
@@ -14,7 +14,7 @@
 // You should have received a copy of the GNU General Public License
 // along with Manta.  If not, see <http://www.gnu.org/licenses/>.
 
-use crate::common::*;
+use crate::integrations_mock::*;
 
 use calamari_runtime::opaque::SessionKeys;
 pub use calamari_runtime::{
